#10073: Merge logit_bw to TTNN
