+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -906,8 +906,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.threshold_bw
 
-.. autofunction:: tt_lib.tensor.logit_bw
-
 .. autofunction:: tt_lib.tensor.logiteps_bw
 
 .. autofunction:: tt_lib.tensor.softsign_bw
