+++ docs/source/ttnn/ttnn/ttnn/logit_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.logit_bw:
+
+ttnn.logit_bw
+#################
+
+.. autofunction:: ttnn.logit_bw
