+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -32,6 +32,7 @@ enum class UnaryBackwardOpType {
     RSQRT_BW,
     NEG_BW,
     RELU_BW,
+    LOGIT_BW,
 };
 
 struct UnaryBackwardFunction{
